// api-tests/generated/models/IssueTypeScreenSchemeProjectAssociation.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* eslint-disable */
/**
 * Associated issue type screen scheme and project.
 */
export type IssueTypeScreenSchemeProjectAssociation = {
  /**
   * The ID of the issue type screen scheme.
   */
  issueTypeScreenSchemeId?: string;
  /**
   * The ID of the project.
   */
  projectId?: string;
};